many: make shell scripts shellcheck-clean
Also make run-checks run shellcheck on the data completion scripts as
well as on tests/lib/*.sh. Includes making pinentry-fake.sh an actual
script (so shellcheck can check it too) instead of a here-document
inside another one.

+++ tests/lib/reset.sh
@@ -2,18 +2,19 @@
 
 set -e -x
 
-. $TESTSLIB/dirs.sh
+# shellcheck source=tests/lib/dirs.sh
+. "$TESTSLIB/dirs.sh"
 
 reset_classic() {
     systemctl stop snapd.service snapd.socket
 
-    sh -x ${SPREAD_PATH}/debian/snapd.postrm purge
+    sh -x "${SPREAD_PATH}/debian/snapd.postrm" purge
     # extra purge
-    rm -rvf /var/snap $SNAPMOUNTDIR/bin
-    mkdir -p $SNAPMOUNTDIR /var/snap /var/lib/snapd
-    if [ "$(find $SNAPMOUNTDIR /var/snap -mindepth 1 -print -quit)" ]; then
+    rm -rvf /var/snap "${SNAPMOUNTDIR:?}/bin"
+    mkdir -p "$SNAPMOUNTDIR" /var/snap /var/lib/snapd
+    if [ "$(find "$SNAPMOUNTDIR" /var/snap -mindepth 1 -print -quit)" ]; then
         echo "postinst purge failed"
-        ls -lR $SNAPMOUNTDIR/ /var/snap/
+        ls -lR "$SNAPMOUNTDIR"/ /var/snap/
         exit 1
     fi
 
@@ -25,13 +26,13 @@ reset_classic() {
     rm -f /tmp/core* /tmp/ubuntu-core*
 
     if [ "$1" = "--reuse-core" ]; then
-        $(cd / && tar xzf $SPREAD_PATH/snapd-state.tar.gz)
-        escaped_snap_mount_dir="$(systemd-escape --path $SNAPMOUNTDIR)"
+        tar -C/ -xzf "$SPREAD_PATH/snapd-state.tar.gz"
+        escaped_snap_mount_dir="$(systemd-escape --path "$SNAPMOUNTDIR")"
         mounts="$(systemctl list-unit-files --full | grep "^$escaped_snap_mount_dir[-.].*\.mount" | cut -f1 -d ' ')"
         services="$(systemctl list-unit-files --full | grep "^$escaped_snap_mount_dir[-.].*\.service" | cut -f1 -d ' ')"
         systemctl daemon-reload # Workaround for http://paste.ubuntu.com/17735820/
         for unit in $mounts $services; do
-            systemctl start $unit
+            systemctl start "$unit"
         done
     fi
 
@@ -45,9 +46,10 @@ reset_classic() {
 
 reset_all_snap() {
     # remove all leftover snaps
+    # shellcheck source=tests/lib/names.sh
     . "$TESTSLIB/names.sh"
 
-    for snap in $SNAPMOUNTDIR/*; do
+    for snap in "$SNAPMOUNTDIR"/*; do
         snap="${snap:6}"
         case "$snap" in
             "bin" | "$gadget_name" | "$kernel_name" | core )
@@ -61,7 +63,7 @@ reset_all_snap() {
     # ensure we have the same state as initially
     systemctl stop snapd.service snapd.socket
     rm -rf /var/lib/snapd/*
-    $(cd / && tar xzf $SPREAD_PATH/snapd-state.tar.gz)
+    tar -C/ -xzf "$SPREAD_PATH/snapd-state.tar.gz"
     rm -rf /root/.snap
     if [ "$1" != "--keep-stopped" ]; then
         systemctl start snapd.service snapd.socket
@@ -75,6 +77,7 @@ else
 fi
 
 if [ "$REMOTE_STORE" = staging ] && [ "$1" = "--store" ]; then
+    # shellcheck source=tests/lib/store.sh
     . $TESTSLIB/store.sh
     teardown_staging_store
 fi
